+++ Webapp/sources/blueprints/manage_workgroup/routes.py
@@ -5,7 +5,7 @@ from flask_login import (  # protects a view function against anonymous users
     current_user,
     login_required,
 )
-from decorators import principal_investigator_required
+from sources.decorators import principal_investigator_required
 from sources import models, services
 from sources.auxiliary import (
     duplicate_notification_check,
